pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "starfold"
version = "0.0.1"
description = "UMAP + Optuna-tuned HDBSCAN + trustworthiness + statistical noise baseline, for any numerical feature matrix."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [
    { name = "Andreas W. Neitzel", email = "andreaswneitzel@gmail.com" },
]
keywords = [
    "clustering",
    "unsupervised-learning",
    "manifold-learning",
    "umap",
    "hdbscan",
    "optuna",
    "trustworthiness",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "scikit-learn>=1.4",
    "umap-learn>=0.5.5",
    "hdbscan>=0.8.33",
    "optuna>=3.6",
    "matplotlib>=3.8",
    "platformdirs>=4.0",
    "joblib>=1.3",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "hypothesis>=6.100",
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.7",
    "nbstripout>=0.7",
    "jupyter>=1.0",
]

[project.urls]
Homepage = "https://github.com/AndreasWNeitzel/starfold"
Paper = "https://doi.org/10.1051/0004-6361/202451718"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
starfold = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = [
    "legacy",
    "docs/tutorial_*.ipynb",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",
    "D213",
    "COM812",
    "ISC001",
    # The package convention (documented in src/.../__init__.py) uses `X` for
    # feature matrices and `X_high`/`X_low` for paired high/low embeddings;
    # these are the names used throughout the scientific literature this
    # tool follows, so PEP 8 N803/N806 do not apply.
    "N803",
    "N806",
    # Small integer literals (``ndim != 2``, ``percentile=99.7``) read better
    # inline than behind a named constant.
    "PLR2004",
    # Many-argument keyword-only APIs are the scientific-Python convention
    # (think StandardScaler, UMAP, TSNE); the "too many arguments" rule would
    # force artificial config objects without readability gain.
    "PLR0913",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "D", "ANN", "PLR2004", "INP001"]
"examples/**" = ["D", "ANN", "T201", "INP001"]
"docs/**" = ["D", "ANN", "INP001"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
strict = true
python_version = "3.11"
files = ["src"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "umap",
    "umap.*",
    "hdbscan",
    "hdbscan.*",
    "sklearn.*",
    "scipy.*",
    "optuna.*",
    "matplotlib.*",
    "joblib.*",
    "platformdirs",
    "tqdm.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/starfold"]
branch = true
omit = ["*/plotting.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 0
